Deduplicate device_key_to_bluetooth_entity_key
