--- homeassistant/components/bluemaestro/device.py
@@ -1,15 +1,0 @@
-"""Support for BlueMaestro devices."""
-from __future__ import annotations
-
-from bluemaestro_ble import DeviceKey
-
-from homeassistant.components.bluetooth.passive_update_processor import (
-    PassiveBluetoothEntityKey,
-)
-
-
-def device_key_to_bluetooth_entity_key(
-    device_key: DeviceKey,
-) -> PassiveBluetoothEntityKey:
-    """Convert a device key to an entity key."""
-    return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
