+++ homeassistant/components/ruuvitag_ble/sensor.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 from typing import Optional, Union
 
 from sensor_state_data import (
-    DeviceKey,
     SensorDeviceClass as SSDSensorDeviceClass,
     SensorUpdate,
     Units,
@@ -14,7 +13,6 @@ from homeassistant import config_entries, const
 from homeassistant.components.bluetooth.passive_update_processor import (
     PassiveBluetoothDataProcessor,
     PassiveBluetoothDataUpdate,
-    PassiveBluetoothEntityKey,
     PassiveBluetoothProcessorCoordinator,
     PassiveBluetoothProcessorEntity,
 )
@@ -25,6 +23,7 @@ from homeassistant.components.sensor import (
     SensorStateClass,
 )
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.bluetooth import device_key_to_bluetooth_entity_key
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.sensor import (
     sensor_description_to_key,
@@ -79,13 +78,6 @@ SENSOR_DESCRIPTIONS = {
 }
 
 
-def _device_key_to_bluetooth_entity_key(
-    device_key: DeviceKey,
-) -> PassiveBluetoothEntityKey:
-    """Convert a device key to an entity key."""
-    return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
-
-
 def sensor_update_to_bluetooth_data_update(
     sensor_update: SensorUpdate,
 ) -> PassiveBluetoothDataUpdate:
@@ -96,18 +88,18 @@ def sensor_update_to_bluetooth_data_update(
             for device_id, device_info in sensor_update.devices.items()
         },
         entity_descriptions={
-            _device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
+            device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
                 sensor_description_to_key(description)
             ]
             for device_key, description in sensor_update.entity_descriptions.items()
             if sensor_description_to_key(description) in SENSOR_DESCRIPTIONS
         },
         entity_data={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
+            device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
             for device_key, sensor_values in sensor_update.entity_values.items()
         },
         entity_names={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.name
+            device_key_to_bluetooth_entity_key(device_key): sensor_values.name
             for device_key, sensor_values in sensor_update.entity_values.items()
         },
     )
