+++ homeassistant/components/sensorpush/sensor.py
@@ -3,13 +3,12 @@ from __future__ import annotations
 
 from typing import Optional, Union
 
-from sensorpush_ble import DeviceClass, DeviceKey, SensorUpdate, Units
+from sensorpush_ble import DeviceClass, SensorUpdate, Units
 
 from homeassistant import config_entries
 from homeassistant.components.bluetooth.passive_update_processor import (
     PassiveBluetoothDataProcessor,
     PassiveBluetoothDataUpdate,
-    PassiveBluetoothEntityKey,
     PassiveBluetoothProcessorCoordinator,
     PassiveBluetoothProcessorEntity,
 )
@@ -26,6 +25,7 @@ from homeassistant.const import (
     TEMP_CELSIUS,
 )
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers.bluetooth import device_key_to_bluetooth_entity_key
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.sensor import (
     sensor_description_to_key,
@@ -66,13 +66,6 @@ SENSOR_DESCRIPTIONS = {
 }
 
 
-def _device_key_to_bluetooth_entity_key(
-    device_key: DeviceKey,
-) -> PassiveBluetoothEntityKey:
-    """Convert a device key to an entity key."""
-    return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
-
-
 def sensor_update_to_bluetooth_data_update(
     sensor_update: SensorUpdate,
 ) -> PassiveBluetoothDataUpdate:
@@ -83,18 +76,18 @@ def sensor_update_to_bluetooth_data_update(
             for device_id, device_info in sensor_update.devices.items()
         },
         entity_descriptions={
-            _device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
+            device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
                 sensor_description_to_key(description)
             ]
             for device_key, description in sensor_update.entity_descriptions.items()
             if description.device_class and description.native_unit_of_measurement
         },
         entity_data={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
+            device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
             for device_key, sensor_values in sensor_update.entity_values.items()
         },
         entity_names={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.name
+            device_key_to_bluetooth_entity_key(device_key): sensor_values.name
             for device_key, sensor_values in sensor_update.entity_values.items()
         },
     )
