+++ homeassistant/helpers/bluetooth.py
@@ -1,0 +1,21 @@
+"""Common functions related to Bluetooth device management."""
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from homeassistant.components.bluetooth.passive_update_processor import (
+    PassiveBluetoothEntityKey,
+)
+
+if TYPE_CHECKING:
+    # `sensor_state_data` is a second-party library (i.e. maintained by Home Assistant
+    # core members) which is not strictly required by Home Assistant.
+    # Therefore, we import it as a type hint only.
+    from sensor_state_data import DeviceKey
+
+
+def device_key_to_bluetooth_entity_key(
+    device_key: DeviceKey,
+) -> PassiveBluetoothEntityKey:
+    """Convert a sensor_state_data device key to an entity key."""
+    return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
